requirements.txt
# Core dependencies
openai>=1.12.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
langchain-text-splitters>=0.0.1
chromadb>=0.4.22
pypdf>=4.0.0
python-dotenv>=1.0.0

# Utilities
tiktoken>=0.5.2
numpy>=1.24.0
pandas>=2.0.0

# Evaluation and monitoring
pytest>=7.4.0

